WIP: fix editor tabs linting errors
