+++ package.json
@@ -25,9 +25,11 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@monaco-editor/react": "^4.6.0",
+    "@types/lodash": "^4.17.7",
     "@vercel/speed-insights": "^1.0.12",
     "contentful-import": "^9.4.58",
     "framer-motion": "^10.16.12",
+    "lodash": "^4.17.21",
     "lottie-react": "^2.4.0",
     "mongodb": "^6.3.0",
     "next": "14.2.7",
@@ -53,7 +55,6 @@
     "@mdx-js/react": "^3.0.1",
     "@next/mdx": "^14.2.11",
     "@stylistic/eslint-plugin-ts": "^2.8.0",
-    "@types/lodash": "^4.17.7",
     "@types/mdx": "^2.0.13",
     "@types/node": "^20",
     "@types/react": "^18",
@@ -80,7 +81,6 @@
     "graphql": "^16.9.0",
     "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
-    "lodash": "^4.17.21",
     "prettier": "^3.3.2",
     "sort-json": "^2.0.1",
     "typescript": "^5"
